Update SecretSantaCreator
Update secret santa algorithm for partner constraint.
A constraints list is added for each participant. Algorithm
iterates through each participant, removing the list of people
that cannot be allocated to participant, shuffle, then allocate
the last on the list to participant.
This person is then removed from the givers list.