Update SecretSantaCreator
Add final constraint of not allowing the same secret santa as
the last year.
Changed .build_constraints algorithm. Previously it plucked
full_name and partner_full_name, now it iterates through each
participant to create an array that includes last years
secret santa. Not very efficient at all but should be ok because
the list is not expected to be in the thousands.